Merge/pull request

See merge request cloud_arv/health-tech/partner-solution/cariva-playground/playground-dashboard!8

+++ src/modules/SetNewPasswordModule/SetNewPasswordForm/index.tsx
@@ -63,13 +63,9 @@ const SetNewPasswordForm = () => {
         apiUrl,
         submitParams,
       )
-      if (response.data) {
-        const { data } = response.data
-        console.log(data)
-      }
       if (response?.status === 204) {
-        setCookie('passwordChanged', true) // necessary
-        if (resetPasswordToken) {
+      setCookie('passwordChanged', true) // necessary
+        if (resetPasswordToken && resetPasswordToken?.length > 0) {
           router.push(webPaths.home)
           deleteCookie('resetPasswordToken')
           setToastOpen(true, {
@@ -149,6 +145,7 @@ const SetNewPasswordForm = () => {
                   fullWidth
                   autoFocus
                   id={fieldname.NEW_PASSWORD}
+                  data-testid={fieldname.NEW_PASSWORD}
                   type={showPassword ? "text" : "password"}
                   placeholder={!newPassword ? "ตั้งค่ารหัสผ่าน" : ""}
                   name={fieldname.NEW_PASSWORD}
@@ -164,6 +161,7 @@ const SetNewPasswordForm = () => {
                       <a>
                         <IconButton
                           aria-label="Toggle show password"
+                          data-testid="button-toggle-show-new-password"
                           onClick={() => setShowPassword((prev) => !prev)}
                         >
                           {showPassword ? (
@@ -209,6 +207,7 @@ const SetNewPasswordForm = () => {
                   name={fieldname.CONFIRM_NEW_PASSWORD}
                   type={showConfirmPassword ? "text" : "password"}
                   id={fieldname.CONFIRM_NEW_PASSWORD}
+                  data-testid={fieldname.CONFIRM_NEW_PASSWORD}
                   placeholder={!confirmNewPassword ? "ตั้งรหัสผ่านใหม่อีกครั้ง" : ""}
                   value={(confirmNewPassword as unknown) ?? ''}
                   onChange={(event) => {
@@ -221,6 +220,7 @@ const SetNewPasswordForm = () => {
                     endAdornment: (
                       <a>
                         <IconButton
+                          data-testid="button-toggle-show-confirm-password"
                           aria-label="Toggle show password"
                           onClick={() => setShowComfirmPassword((prev) => !prev)}
                         >
@@ -274,6 +274,7 @@ const SetNewPasswordForm = () => {
             </Typography>
           </Stack>
           <SubmitButtonStyle
+            data-testid='button-set-new-password'
             type="submit"
             disabled={isDisableSubmit}
           >
